// src/types.ts
export interface Event {
  id: string;
  slug: string;
  title: string;
  date: string;
  location: string;
  category: string;
  image?: string;
  imagecredit?: string;
  gallery?: string[];
  video?: string;
  excerpt: string;
  content: string;
  status: 'draft' | 'published';
  scheduledat?: string;
  price?: number;
  ispaid?: boolean;
  maxseats?: number;
  bookingscount?: number;
}